According to the MIR author statement, MIR does not generate position-independent code. This means that the code can only be executed in the location where it was originally generated. Since it is not possible to move the generated code to a specific memory location, we need to refactor a code cache that solely stores the memory address of the generated code.

The following statistics demonstrate the performance improvement observed when importing JIT (Just-in-Time) compilation on various host machines.

* Core i7-8700

|Metric| rv32emu (interpreter) |rv32emu (JIT) |Speedup|
|--------| -------- | -------- | -------- | 
| CoreMark | 1155.174 (iter/s)| 1932.496 (iter/s) |+67.2％|
| dhrystone  | 1282 DMIPS |2495.11 DMIPS |+94.6%|

* eMAG 8180

|Metric| rv32emu (interpreter) |rv32emu (baseline JIT) |Speedup|
|--------| -------- | -------- | -------- | 
| CoreMark | 438.073 (iter/s)| 644.426 (iter/s) |+47.1％|
| dhrystone  | 354.60 DMIPS |867.00 DMIPS |+144%|