@jserv 2023-06-18
Known issue: JIT works on Linux/Arm64, but it fails on macOS/Arm64 (Apple M1).
```
$ make ENABLE_JIT=1 check
Running hello.elf ... [OK]
Running puzzle.elf ... Failed.
make: *** [check] Error 1
```