@jserv 2023-06-18
Known issue: misalign test case fails.
```shell
$ make ENABLE_JIT=1 misalign 
Segmentation fault
Running aes.elf ... Failed.
```